learn-gemini-model/pyproject.toml
[project]
name = "learn-gemini-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.73.0",
    "httpx[socks]>=0.28.1",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "tiktoken>=0.12.0",
]
